+++ azurelinuxagent/common/protocol/wire.py
@@ -20,20 +20,19 @@ import datetime
 import json
 import os
 import random
-import re
 import time
 import xml.sax.saxutils as saxutils
 from datetime import datetime
-from threading import Lock
 
 import azurelinuxagent.common.conf as conf
 import azurelinuxagent.common.logger as logger
 import azurelinuxagent.common.utils.textutil as textutil
-from azurelinuxagent.common.datacontract import validate_param, set_properties
-from azurelinuxagent.common.event import add_periodic, WALAEventOperation, CONTAINER_ID_ENV_VARIABLE
+from azurelinuxagent.common.datacontract import validate_param
+from azurelinuxagent.common.event import add_periodic, WALAEventOperation
 from azurelinuxagent.common.exception import ProtocolNotFoundError, \
     ResourceGoneError, ExtensionDownloadError, InvalidContainerError, ProtocolError, HttpError
 from azurelinuxagent.common.future import httpclient, bytebuffer
+from azurelinuxagent.common.goal_state import GoalState, TRANSPORT_CERT_FILE_NAME, TRANSPORT_PRV_FILE_NAME
 from azurelinuxagent.common.protocol.hostplugin import HostPluginProtocol
 from azurelinuxagent.common.protocol.restapi import *
 from azurelinuxagent.common.telemetryevent import TelemetryEventList
@@ -41,11 +40,10 @@ from azurelinuxagent.common.utils import fileutil, restutil
 from azurelinuxagent.common.utils.archive import StateFlusher
 from azurelinuxagent.common.utils.cryptutil import CryptUtil
 from azurelinuxagent.common.utils.textutil import parse_doc, findall, find, \
-    findtext, getattrib, gettext, remove_bom, get_bytes_from_pem, parse_json
+    findtext, gettext, remove_bom, get_bytes_from_pem, parse_json
 from azurelinuxagent.common.version import AGENT_NAME, CURRENT_VERSION
 
 VERSION_INFO_URI = "http://{0}/?comp=versions"
-GOAL_STATE_URI = "http://{0}/machine/?comp=goalstate"
 HEALTH_REPORT_URI = "http://{0}/machine?comp=health"
 ROLE_PROP_URI = "http://{0}/machine?comp=roleProperties"
 TELEMETRY_URI = "http://{0}/machine?comp=telemetrydata"
@@ -57,14 +55,8 @@ HOSTING_ENV_FILE_NAME = "HostingEnvironmentConfig.xml"
 SHARED_CONF_FILE_NAME = "SharedConfig.xml"
 CERTS_FILE_NAME = "Certificates.xml"
 REMOTE_ACCESS_FILE_NAME = "RemoteAccess.{0}.xml"
-P7M_FILE_NAME = "Certificates.p7m"
-PEM_FILE_NAME = "Certificates.pem"
 EXT_CONF_FILE_NAME = "ExtensionsConfig.{0}.xml"
 MANIFEST_FILE_NAME = "{0}.{1}.manifest.xml"
-TRANSPORT_CERT_FILE_NAME = "TransportCert.pem"
-TRANSPORT_PRV_FILE_NAME = "TransportPrivate.pem"
-# Store the last retrieved container id as an environment variable to be shared between threads for telemetry purposes
-CONTAINER_ID_ENV_VARIABLE = "AZURE_GUEST_AGENT_CONTAINER_ID"
 
 PROTOCOL_VERSION = "2012-11-30"
 ENDPOINT_FINE_NAME = "WireServer"
@@ -147,8 +139,7 @@ class WireProtocol(Protocol):
 
     def get_ext_handler_pkgs(self, ext_handler):
         logger.verbose("Get extension handler package")
-        goal_state = self.client.get_goal_state()
-        man = self.client.get_ext_manifest(ext_handler, goal_state)
+        man = self.client.get_ext_manifest(ext_handler)
         return man.pkg_list
 
     def get_artifacts_profile(self):
@@ -694,60 +685,6 @@ class WireClient(object):
                 raise
         return resp
 
-    def update_hosting_env(self, goal_state):
-        if goal_state.hosting_env_uri is None:
-            raise ProtocolError("HostingEnvironmentConfig uri is empty")
-        local_file = os.path.join(conf.get_lib_dir(), HOSTING_ENV_FILE_NAME)
-        xml_text = self.fetch_config(goal_state.hosting_env_uri,
-                                     self.get_header())
-        self.save_cache(local_file, xml_text)
-        self._set_hosting_env(HostingEnv(xml_text))
-
-    def update_shared_conf(self, goal_state):
-        if goal_state.shared_conf_uri is None:
-            raise ProtocolError("SharedConfig uri is empty")
-        local_file = os.path.join(conf.get_lib_dir(), SHARED_CONF_FILE_NAME)
-        xml_text = self.fetch_config(goal_state.shared_conf_uri,
-                                     self.get_header())
-        self.save_cache(local_file, xml_text)
-        self._set_shared_conf(SharedConfig(xml_text))
-
-    def update_certs(self, goal_state):
-        if goal_state.certs_uri is None:
-            return
-        local_file = os.path.join(conf.get_lib_dir(), CERTS_FILE_NAME)
-        xml_text = self.fetch_config(goal_state.certs_uri,
-                                     self.get_header_for_cert())
-        self.save_cache(local_file, xml_text)
-        self._set_certs(Certificates(self, xml_text))
-
-    def update_remote_access_conf(self, goal_state):
-        if goal_state.remote_access_uri is None:
-            # Nothing in accounts data.  Just return, nothing to do.
-            return
-        xml_text = self.fetch_config(goal_state.remote_access_uri,
-                                     self.get_header_for_cert())
-        remote_access = RemoteAccess(xml_text)
-        local_file = os.path.join(conf.get_lib_dir(), REMOTE_ACCESS_FILE_NAME.format(remote_access.incarnation))
-        self.save_cache(local_file, xml_text)
-        self._set_remote_access(remote_access)
-
-    def get_remote_access(self):
-        # This property can be None in the GoalState, so not raising a ProtocolError here
-        return self._remote_access
-
-    def update_ext_conf(self, goal_state):
-        if goal_state.ext_uri is None:
-            logger.info("ExtensionsConfig.xml uri is empty")
-            self._set_ext_conf(ExtensionsConfig(None))
-            return
-        incarnation = goal_state.incarnation
-        local_file = os.path.join(conf.get_lib_dir(),
-                                  EXT_CONF_FILE_NAME.format(incarnation))
-        xml_text = self.fetch_config(goal_state.ext_uri, self.get_header())
-        self.save_cache(local_file, xml_text)
-        self._set_ext_conf(ExtensionsConfig(xml_text))
-
     # Type of update performed by _update_from_goal_state()
     class _UpdateType(object):
         # Update the Host GA Plugin client (Container ID and RoleConfigName)
@@ -774,14 +711,11 @@ class WireClient(object):
         """
         Fetches a new goal state and updates the internal state of the WireClient according to the requested 'refresh_type'
         """
-        uri = GOAL_STATE_URI.format(self.get_endpoint())
-
         max_retry = 3
 
         for retry in range(1, max_retry + 1):
             try:
-                new_goal_state_xml = self.fetch_config(uri, self.get_header())
-                new_goal_state = GoalState(new_goal_state_xml)
+                new_goal_state = GoalState(self)
 
                 def update_host_plugin():
                     if self._host_plugin is not None:
@@ -794,20 +728,10 @@ class WireClient(object):
 
                 if self._goal_state is None or refresh_type == WireClient._UpdateType.GoalStateForced or \
                         new_goal_state.incarnation != self._goal_state.incarnation:
-                    def save_goal_state(incarnation, xml_text):
-                        file_name = GOAL_STATE_FILE_NAME.format(incarnation)
-                        goal_state_file = os.path.join(conf.get_lib_dir(), file_name)
-                        self.save_cache(goal_state_file, xml_text)
 
-                    self.goal_state_flusher.flush(datetime.utcnow())
-                    self._set_goal_state(new_goal_state)
-                    save_goal_state(new_goal_state.incarnation, new_goal_state_xml)
-                    self.update_hosting_env(new_goal_state)
-                    self.update_shared_conf(new_goal_state)
-                    self.update_certs(new_goal_state)
-                    self.update_ext_conf(new_goal_state)
-                    self.update_remote_access_conf(new_goal_state)
-                    self.save_cache(os.path.join(conf.get_lib_dir(), INCARNATION_FILE_NAME), new_goal_state.incarnation)
+                    new_goal_state.fetch_full_goal_state(self)
+                    self._goal_state = new_goal_state
+                    self._save_goal_state()
 
                     update_host_plugin()
                 return
@@ -826,35 +750,37 @@ class WireClient(object):
 
         raise ProtocolError("Exceeded max retry updating goal state")
 
-    def _set_goal_state(self, new_goal_state):
-        if new_goal_state is None:
-            logger.warn("Setting empty Goal State object!")
-        self._goal_state = new_goal_state
+    def _save_goal_state(self):
+        try:
+            self.goal_state_flusher.flush(datetime.utcnow())
 
-    def _set_hosting_env(self, new_hosting_env):
-        if new_hosting_env is None:
-            logger.warn("Setting empty Hosting Environment object!")
-        self._hosting_env = new_hosting_env
+        except Exception as e:
+            logger.warn("Failed to save the previous goal state to the history folder: {0}", ustr(e))
 
-    def _set_shared_conf(self, new_shared_config):
-        if new_shared_config is None:
-            logger.warn("Setting empty Shared Config object!")
-        self._shared_conf = new_shared_config
+        try:
+            local_file = os.path.join(conf.get_lib_dir(), INCARNATION_FILE_NAME)
+            self.save_cache(local_file, self._goal_state.incarnation)
 
-    def _set_certs(self, new_certs):
-        if new_certs is None:
-            logger.warn("Setting empty Certs object!")
-        self._certs = new_certs
+            local_file = os.path.join(conf.get_lib_dir(), GOAL_STATE_FILE_NAME.format(self._goal_state.incarnation))
+            self.save_cache(local_file, self._goal_state.xml_text)
 
-    def _set_ext_conf(self, new_ext_conf):
-        if new_ext_conf is None:
-            logger.warn("Setting empty Extension Config object!")
-        self._ext_conf = new_ext_conf
+            local_file = os.path.join(conf.get_lib_dir(), HOSTING_ENV_FILE_NAME)
+            self.save_cache(local_file, self._goal_state.hosting_env.xml_text)
 
-    def _set_remote_access(self, new_remote_access):
-        if new_remote_access is None:
-            logger.warn("Setting empty Remote Access object!")
-        self._remote_access = new_remote_access
+            local_file = os.path.join(conf.get_lib_dir(), SHARED_CONF_FILE_NAME)
+            self.save_cache(local_file, self._goal_state.shared_conf.xml_text)
+
+            local_file = os.path.join(conf.get_lib_dir(), CERTS_FILE_NAME)
+            self.save_cache(local_file, self._goal_state.certs.xml_text)
+
+            local_file = os.path.join(conf.get_lib_dir(), EXT_CONF_FILE_NAME.format(self._goal_state.incarnation))
+            self.save_cache(local_file, self._goal_state.ext_conf.xml_text)
+
+            local_file = os.path.join(conf.get_lib_dir(), REMOTE_ACCESS_FILE_NAME.format(self._goal_state.incarnation))
+            self.save_cache(local_file, self._goal_state.remote_access.xml_text)
+
+        except Exception as e:
+            logger.warn("Failed to save the new goal state to disk: {0}", ustr(e))
 
     def _set_host_plugin(self, new_host_plugin):
         if new_host_plugin is None:
@@ -867,29 +793,29 @@ class WireClient(object):
         return self._goal_state
 
     def get_hosting_env(self):
-        if self._hosting_env is None:
+        if self._goal_state is None:
             raise ProtocolError("Trying to fetch Hosting Environment before initialization!")
-        return self._hosting_env
+        return self._goal_state.hosting_env
 
     def get_shared_conf(self):
-        if self._shared_conf is None:
+        if self._goal_state is None:
             raise ProtocolError("Trying to fetch Shared Conf before initialization!")
-        return self._shared_conf
+        return self._goal_state.shared_conf
 
     def get_certs(self):
         # This property can be None in the GoalState, so not returning a ProtocolError here
-        return self._certs
+        return self._goal_state.certs
 
     def get_ext_conf(self):
-        if self._ext_conf is None:
-            if self.get_goal_state().ext_uri is None:
-                self._set_ext_conf(ExtensionsConfig(None))
-            else:
-                raise ProtocolError("Trying to fetch Extension Conf before initialization!")
-        return self._ext_conf
+        if self._goal_state is None:
+            raise ProtocolError("Trying to fetch Extension Conf before initialization!")
+        return self._goal_state.ext_conf
 
-    def get_ext_manifest(self, ext_handler, goal_state):
-        local_file = MANIFEST_FILE_NAME.format(ext_handler.name, goal_state.incarnation)
+    def get_ext_manifest(self, ext_handler):
+        if self._goal_state is None:
+            raise ProtocolError("Trying to fetch Extension Manifest before initialization!")
+
+        local_file = MANIFEST_FILE_NAME.format(ext_handler.name, self.client.get_goal_state().incarnation)
         local_file = os.path.join(conf.get_lib_dir(), local_file)
 
         try:
@@ -898,6 +824,11 @@ class WireClient(object):
             return ExtensionManifest(xml_text)
         except Exception as e:
             raise ExtensionDownloadError("Failed to retrieve extension manifest. Error: {0}".format(ustr(e)))
+
+    def get_remote_access(self):
+        if self._goal_state is None:
+            raise ProtocolError("Trying to fetch Remote Access before initialization!")
+        return self._goal_state.remote_access
 
     def get_gafamily_manifest(self, vmagent_manifest, goal_state):
         local_file = MANIFEST_FILE_NAME.format(vmagent_manifest.family, goal_state.incarnation)
@@ -1283,365 +1214,6 @@ class VersionInfo(object):
 
     def get_supported(self):
         return self.supported
-
-
-class GoalState(object):
-    def __init__(self, xml_text):
-        if xml_text is None:
-            raise ValueError("GoalState.xml is None")
-        logger.verbose("Load GoalState.xml")
-        self.incarnation = None
-        self.expected_state = None
-        self.hosting_env_uri = None
-        self.shared_conf_uri = None
-        self.remote_access_uri = None
-        self.certs_uri = None
-        self.ext_uri = None
-        self.role_instance_id = None
-        self.role_config_name = None
-        self.container_id = None
-        self.load_balancer_probe_port = None
-        self.xml_text = None
-        self._parse(xml_text)
-
-    def _parse(self, xml_text):
-        self.xml_text = xml_text
-        xml_doc = parse_doc(xml_text)
-        self.incarnation = findtext(xml_doc, "Incarnation")
-        self.expected_state = findtext(xml_doc, "ExpectedState")
-        self.hosting_env_uri = findtext(xml_doc, "HostingEnvironmentConfig")
-        self.shared_conf_uri = findtext(xml_doc, "SharedConfig")
-        self.certs_uri = findtext(xml_doc, "Certificates")
-        self.ext_uri = findtext(xml_doc, "ExtensionsConfig")
-        role_instance = find(xml_doc, "RoleInstance")
-        self.role_instance_id = findtext(role_instance, "InstanceId")
-        role_config = find(role_instance, "Configuration")
-        self.role_config_name = findtext(role_config, "ConfigName")
-        container = find(xml_doc, "Container")
-        self.container_id = findtext(container, "ContainerId")
-        os.environ[CONTAINER_ID_ENV_VARIABLE] = self.container_id
-        self.remote_access_uri = findtext(container, "RemoteAccessInfo")
-        lbprobe_ports = find(xml_doc, "LBProbePorts")
-        self.load_balancer_probe_port = findtext(lbprobe_ports, "Port")
-
-
-class HostingEnv(object):
-    def __init__(self, xml_text):
-        if xml_text is None:
-            raise ValueError("HostingEnvironmentConfig.xml is None")
-        logger.verbose("Load HostingEnvironmentConfig.xml")
-        self.vm_name = None
-        self.role_name = None
-        self.deployment_name = None
-        self.xml_text = None
-        self._parse(xml_text)
-
-    def _parse(self, xml_text):
-        self.xml_text = xml_text
-        xml_doc = parse_doc(xml_text)
-        incarnation = find(xml_doc, "Incarnation")
-        self.vm_name = getattrib(incarnation, "instance")
-        role = find(xml_doc, "Role")
-        self.role_name = getattrib(role, "name")
-        deployment = find(xml_doc, "Deployment")
-        self.deployment_name = getattrib(deployment, "name")
-
-
-class SharedConfig(object):
-    def __init__(self, xml_text):
-        logger.verbose("Load SharedConfig.xml")
-        self.xml_text = xml_text
-
-
-class RemoteAccess(object):
-    """
-    Object containing information about user accounts
-    """
-
-    #
-    # <RemoteAccess>
-    #   <Version/>
-    #   <Incarnation/>
-    #    <Users>
-    #       <User>
-    #         <Name/>
-    #         <Password/>
-    #         <Expiration/>
-    #       </User>
-    #     </Users>
-    #   </RemoteAccess>
-    #
-
-    def __init__(self, xml_text):
-        logger.verbose("Load RemoteAccess.xml")
-        self.version = None
-        self.incarnation = None
-        self.user_list = RemoteAccessUsersList()
-
-        self.xml_text = None
-        self._parse(xml_text)
-
-    def _parse(self, xml_text):
-        """
-        Parse xml document containing user account information
-        """
-        if xml_text is None or len(xml_text) == 0:
-            return None
-        self.xml_text = xml_text
-        xml_doc = parse_doc(xml_text)
-        self.incarnation = findtext(xml_doc, "Incarnation")
-        self.version = findtext(xml_doc, "Version")
-        user_collection = find(xml_doc, "Users")
-        users = findall(user_collection, "User")
-
-        for user in users:
-            remote_access_user = self._parse_user(user)
-            self.user_list.users.append(remote_access_user)
-
-    def _parse_user(self, user):
-        name = findtext(user, "Name")
-        encrypted_password = findtext(user, "Password")
-        expiration = findtext(user, "Expiration")
-        remote_access_user = RemoteAccessUser(name, encrypted_password, expiration)
-        return remote_access_user
-
-
-class UserAccount(object):
-    """
-    Stores information about single user account
-    """
-
-    def __init__(self):
-        self.Name = None
-        self.EncryptedPassword = None
-        self.Password = None
-        self.Expiration = None
-        self.Groups = []
-
-
-class Certificates(object):
-    """
-    Object containing certificates of host and provisioned user.
-    """
-
-    def __init__(self, client, xml_text):
-        logger.verbose("Load Certificates.xml")
-        self.client = client
-        self.cert_list = CertList()
-        self._parse(xml_text)
-
-    def _parse(self, xml_text):
-        """
-        Parse multiple certificates into seperate files.
-        """
-        xml_doc = parse_doc(xml_text)
-        data = findtext(xml_doc, "Data")
-        if data is None:
-            return
-
-        # if the certificates format is not Pkcs7BlobWithPfxContents do not parse it
-        certificateFormat = findtext(xml_doc, "Format")
-        if certificateFormat and certificateFormat != "Pkcs7BlobWithPfxContents":
-            logger.warn("The Format is not Pkcs7BlobWithPfxContents. Format is " + certificateFormat)
-            return
-
-        cryptutil = CryptUtil(conf.get_openssl_cmd())
-        p7m_file = os.path.join(conf.get_lib_dir(), P7M_FILE_NAME)
-        p7m = ("MIME-Version:1.0\n"
-               "Content-Disposition: attachment; filename=\"{0}\"\n"
-               "Content-Type: application/x-pkcs7-mime; name=\"{1}\"\n"
-               "Content-Transfer-Encoding: base64\n"
-               "\n"
-               "{2}").format(p7m_file, p7m_file, data)
-
-        self.client.save_cache(p7m_file, p7m)
-
-        trans_prv_file = os.path.join(conf.get_lib_dir(),
-                                      TRANSPORT_PRV_FILE_NAME)
-        trans_cert_file = os.path.join(conf.get_lib_dir(),
-                                       TRANSPORT_CERT_FILE_NAME)
-        pem_file = os.path.join(conf.get_lib_dir(), PEM_FILE_NAME)
-        # decrypt certificates
-        cryptutil.decrypt_p7m(p7m_file, trans_prv_file, trans_cert_file,
-                              pem_file)
-
-        # The parsing process use public key to match prv and crt.
-        buf = []
-        begin_crt = False
-        begin_prv = False
-        prvs = {}
-        thumbprints = {}
-        index = 0
-        v1_cert_list = []
-        with open(pem_file) as pem:
-            for line in pem.readlines():
-                buf.append(line)
-                if re.match(r'[-]+BEGIN.*KEY[-]+', line):
-                    begin_prv = True
-                elif re.match(r'[-]+BEGIN.*CERTIFICATE[-]+', line):
-                    begin_crt = True
-                elif re.match(r'[-]+END.*KEY[-]+', line):
-                    tmp_file = self.write_to_tmp_file(index, 'prv', buf)
-                    pub = cryptutil.get_pubkey_from_prv(tmp_file)
-                    prvs[pub] = tmp_file
-                    buf = []
-                    index += 1
-                    begin_prv = False
-                elif re.match(r'[-]+END.*CERTIFICATE[-]+', line):
-                    tmp_file = self.write_to_tmp_file(index, 'crt', buf)
-                    pub = cryptutil.get_pubkey_from_crt(tmp_file)
-                    thumbprint = cryptutil.get_thumbprint_from_crt(tmp_file)
-                    thumbprints[pub] = thumbprint
-                    # Rename crt with thumbprint as the file name
-                    crt = "{0}.crt".format(thumbprint)
-                    v1_cert_list.append({
-                        "name": None,
-                        "thumbprint": thumbprint
-                    })
-                    os.rename(tmp_file, os.path.join(conf.get_lib_dir(), crt))
-                    buf = []
-                    index += 1
-                    begin_crt = False
-
-        # Rename prv key with thumbprint as the file name
-        for pubkey in prvs:
-            thumbprint = thumbprints[pubkey]
-            if thumbprint:
-                tmp_file = prvs[pubkey]
-                prv = "{0}.prv".format(thumbprint)
-                os.rename(tmp_file, os.path.join(conf.get_lib_dir(), prv))
-                logger.info("Found private key matching thumbprint {0}".format(thumbprint))
-            else:
-                # Since private key has *no* matching certificate,
-                # it will not be named correctly
-                logger.warn("Found NO matching cert/thumbprint for private key!")
-
-        # Log if any certificates were found without matching private keys
-        # This can happen (rarely), and is useful to know for debugging
-        for pubkey in thumbprints:
-            if not pubkey in prvs:
-                msg = "Certificate with thumbprint {0} has no matching private key."
-                logger.info(msg.format(thumbprints[pubkey]))
-
-        for v1_cert in v1_cert_list:
-            cert = Cert()
-            set_properties("certs", cert, v1_cert)
-            self.cert_list.certificates.append(cert)
-
-    def write_to_tmp_file(self, index, suffix, buf):
-        file_name = os.path.join(conf.get_lib_dir(),
-                                 "{0}.{1}".format(index, suffix))
-        self.client.save_cache(file_name, "".join(buf))
-        return file_name
-
-
-class ExtensionsConfig(object):
-    def __init__(self, xml_text):
-        logger.verbose("Load ExtensionsConfig.xml")
-        self.ext_handlers = ExtHandlerList()
-        self.vmagent_manifests = VMAgentManifestList()
-        self.status_upload_blob = None
-        self.status_upload_blob_type = None
-        self.artifacts_profile_blob = None
-        if xml_text is not None:
-            self._parse(xml_text)
-
-    def _parse(self, xml_text):
-        xml_doc = parse_doc(xml_text)
-
-        ga_families_list = find(xml_doc, "GAFamilies")
-        ga_families = findall(ga_families_list, "GAFamily")
-
-        for ga_family in ga_families:
-            family = findtext(ga_family, "Name")
-            uris_list = find(ga_family, "Uris")
-            uris = findall(uris_list, "Uri")
-            manifest = VMAgentManifest()
-            manifest.family = family
-            for uri in uris:
-                manifestUri = VMAgentManifestUri(uri=gettext(uri))
-                manifest.versionsManifestUris.append(manifestUri)
-            self.vmagent_manifests.vmAgentManifests.append(manifest)
-
-        plugins_list = find(xml_doc, "Plugins")
-        plugins = findall(plugins_list, "Plugin")
-        plugin_settings_list = find(xml_doc, "PluginSettings")
-        plugin_settings = findall(plugin_settings_list, "Plugin")
-
-        for plugin in plugins:
-            ext_handler = self._parse_plugin(plugin)
-            self.ext_handlers.extHandlers.append(ext_handler)
-            self._parse_plugin_settings(ext_handler, plugin_settings)
-
-        self.status_upload_blob = findtext(xml_doc, "StatusUploadBlob")
-        self.artifacts_profile_blob = findtext(xml_doc, "InVMArtifactsProfileBlob")
-
-        status_upload_node = find(xml_doc, "StatusUploadBlob")
-        self.status_upload_blob_type = getattrib(status_upload_node,
-                                                 "statusBlobType")
-        logger.verbose("Extension config shows status blob type as [{0}]",
-                       self.status_upload_blob_type)
-
-    def _parse_plugin(self, plugin):
-        ext_handler = ExtHandler()
-        ext_handler.name = getattrib(plugin, "name")
-        ext_handler.properties.version = getattrib(plugin, "version")
-        ext_handler.properties.state = getattrib(plugin, "state")
-
-        location = getattrib(plugin, "location")
-        failover_location = getattrib(plugin, "failoverlocation")
-        for uri in [location, failover_location]:
-            version_uri = ExtHandlerVersionUri()
-            version_uri.uri = uri
-            ext_handler.versionUris.append(version_uri)
-        return ext_handler
-
-    def _parse_plugin_settings(self, ext_handler, plugin_settings):
-        if plugin_settings is None:
-            return
-
-        name = ext_handler.name
-        version = ext_handler.properties.version
-        settings = [x for x in plugin_settings \
-                    if getattrib(x, "name") == name and \
-                    getattrib(x, "version") == version]
-
-        if settings is None or len(settings) == 0:
-            return
-
-        runtime_settings = None
-        runtime_settings_node = find(settings[0], "RuntimeSettings")
-        seqNo = getattrib(runtime_settings_node, "seqNo")
-        runtime_settings_str = gettext(runtime_settings_node)
-        try:
-            runtime_settings = json.loads(runtime_settings_str)
-        except ValueError as e:
-            logger.error("Invalid extension settings")
-            return
-
-        depends_on_level = 0
-        depends_on_node = find(settings[0], "DependsOn")
-        if depends_on_node != None:
-            try:
-                depends_on_level = int(getattrib(depends_on_node, "dependencyLevel"))
-            except (ValueError, TypeError):
-                logger.warn("Could not parse dependencyLevel for handler {0}. Setting it to 0".format(name))
-                depends_on_level = 0
-
-        for plugin_settings_list in runtime_settings["runtimeSettings"]:
-            handler_settings = plugin_settings_list["handlerSettings"]
-            ext = Extension()
-            # There is no "extension name" in wire protocol.
-            # Put
-            ext.name = ext_handler.name
-            ext.sequenceNumber = seqNo
-            ext.publicSettings = handler_settings.get("publicSettings")
-            ext.protectedSettings = handler_settings.get("protectedSettings")
-            ext.dependencyLevel = depends_on_level
-            thumbprint = handler_settings.get(
-                "protectedSettingsCertThumbprint")
-            ext.certificateThumbprint = thumbprint
-            ext_handler.properties.extensions.append(ext)
 
 
 class ExtensionManifest(object):
